usbuart: Moved aux_serial_init() into aux_serial.c
